add first test for verion endpoint wip
